Rawr3: A whole bunch of little UI fixes * <> = Available for Optimizer in non-scrolling header [Fixed] * Show Welcome only once [Fixed] * Buffs > Elixirs and Flasks [Fixed] * Buffs > Potions [Fixed] * Edit Gemming Templates, Remove 2 spaces between Blue and Prismatic [Fixed] * You Stand Behind Boss, Didn't Check [Fixed], Needs Perc convertor [Fixed] * Attacks, Didn't Check [Fixed], Need to set the Special bosses to have an IsDefaultMelee [Fixed], There's enough room to increase font size to 8 from 7 [Fixed], Need to set format limits on the strings for Attack.ToString() [Fixed] * Impedences, None of these checked [Fixed] * MultiTargs, Didn't Check [Fixed], Needs Perc Convertor [Fixed]
git-tfs-id: [https://tfs.codeplex.com/tfs/TFS03]$/Rawr;C52066

+++ Rawr.Enhance/CalculationOptionsEnhance.cs
@@ -101,7 +101,7 @@ namespace Rawr.Enhance
             _targetNatureResistance = (float)boss.Resistance(ItemDamageType.Nature);
             _multipleTargets = boss.MaxNumTargets > 1;
             _additionalTargets = (int)boss.MaxNumTargets - 1;
-            _additionalTargetPercent = boss.MultiTargsPerc;
+            _additionalTargetPercent = (float)boss.MultiTargsPerc;
         }
 
         public bool PriorityInUse(EnhanceAbility abilityType)
